passenger make trip

+++ app/src/main/java/com/israa/atmodrive/home/data/models/ConfirmTripResponse.kt
@@ -3,5 +3,5 @@ package com.israa.atmodrive.home.data.models
 data class ConfirmTripResponse(
     val message: String,
     val status: Boolean,
-    val trip_id: Int?
+    val trip_id: Long?
 )
